Coding Standards: Cast `wp_count_terms()` result to `int` before using in `ceil()`.
This addresses two instances of the (numeric string) return value from `wp_count_terms()` being used directly in `ceil()`, which expects an `int|float`.

Affected methods:
* `WP_Sitemaps_Taxonomies::get_max_num_pages()`
* `wp_nav_menu_item_taxonomy_meta_box()`

Reference: [https://www.php.net/manual/en/function.ceil.php PHP Manual: ceil()].

Follow-up to [14248], [14291], [14569], [14943], [48072], [57648].

Props justlevine.
See #52217.
Built from https://develop.svn.wordpress.org/trunk@59462


git-svn-id: http://core.svn.wordpress.org/trunk@58848 1a063a9b-81f0-0310-95a4-ce76da25c4cd

+++ wp-admin/includes/nav-menu.php
@@ -875,7 +875,7 @@ function wp_nav_menu_item_taxonomy_meta_box( $data_object, $box ) {
 	}
 
 	$num_pages = (int) ceil(
-		wp_count_terms(
+		(int) wp_count_terms(
 			array_merge(
 				$args,
 				array(
